Properly cast Nselected to unsigned int before multiplying.

+++ src/CoordCovarMatrix.cpp
@@ -63,7 +63,7 @@ void CoordCovarMatrix::get_frame_coords(Darray& arrayIn, Frame const& frameIn, A
 const
 {
   arrayIn.clear();
-  arrayIn.reserve( maskIn.Nselected() * nelt_ );
+  arrayIn.reserve( (unsigned int)maskIn.Nselected() * nelt_ );
   for (AtomMask::const_iterator at = maskIn.begin(); at != maskIn.end(); ++at) {
     const double* XYZ = frameIn.XYZ(*at);
     arrayIn.push_back( XYZ[0] );
